Implement PastOrder and ReOrder Functionality

+++ app/controllers/customers/restaurants_controller.rb
@@ -1,6 +1,11 @@
 class Customers::RestaurantsController < ApplicationController
   def show
-    session.delete(:order_id)
     @restaurant = Restaurant.find(params[:id])
+    # byebug
+    if params.has_key?(:order_id)
+      @order = current_customer.orders.find(params[:order_id])
+    else
+      session.delete(:order_id)
+    end
   end
 end
